refactor(microfloat): split the block formats into core / manipulators / iostream (#1334)
Phase 2 group 5 of the header layering: the MX / NVFP / ZFP block formats and the two
element types they are built from.

                  umbrella before        core after
    microfloat    61,980 lines / 5 I/O   44,567 / 0
    e8m0          61,160 / 4             43,857 / 0
    mxfloat       73,781 / 5             56,069 / 0
    nvblock       73,165 / 5             55,588 / 0
    zfpblock      77,971 / 5             55,705 / 0

Dependency order drove the work, as it did for the cascades. mxfloat and nvblock included
the UMBRELLAS of their element and scale types -- microfloat and e8m0 -- so those two had
to be layered first. The block cores now take microfloat/core.hpp and e8m0/core.hpp, and
the block iostream.hpp layers take the corresponding stream headers. That is what lets a
block-format core reach zero instead of inheriting five I/O headers through its element
type.

This group was cheaper than the cascades for two reasons worth recording. There are ZERO
cout/cerr sites across all five impls, so no fprintf conversions and none of the discarded
`if constexpr` / std::cout declaration trap that dd hit. And zfpblock needed almost
nothing: its text already lived in manipulators.hpp, it defines no stream operator, and its
<iomanip> was a dead include that nothing in the file used -- so it gets a core.hpp and an
umbrella rewrite but no iostream.hpp.

One real break, found by building the consumers: static/block/mxfloat/api/api.cpp needs
to_binary(e8m0). Pointing mxblock_impl.hpp at e8m0/core.hpp meant mxfloat's manipulators
layer had to take e8m0/manipulators.hpp alongside microfloat's, because the block streams
both its scale and its elements.

Closed 20 IWYU gaps across the five subtrees: <iomanip> for std::setw in three
attributes.hpp, <string> in five exceptions.hpp, and the <cstddef>/<cstdint>/<type_traits>
the manipulators and impls had been getting transitively. Sweep clean at 0 gaps across all
46 headers.

Verified: all five core gates pass on gcc and clang with #error tripwires on the stream
include guards; 41 block-format consumers build and run clean on both compilers; ASCII
clean.

A note on one non-finding: tools/ucalc/ucalc.cpp timed out in the sweep, but it is a REPL
and my loop did not close stdin. With </dev/null it exits 0 on both compilers and on main.
That is unlike education/interactive/expansion_algebra, which genuinely loops forever on
EOF -- the two look identical in a sweep and only one is a defect.

Part of #1334.
