+++ include/sw/universal/number/e8m0/attributes.hpp
@@ -5,6 +5,7 @@
 // SPDX-License-Identifier: MIT
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <iomanip>       // std::setw
 #include <cstdint>
 #include <string>
 #include <sstream>
