+++ include/sw/universal/number/e8m0/manipulators.hpp
@@ -5,6 +5,15 @@
 // SPDX-License-Identifier: MIT
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <cstdint>       // the fixed-width integer types
+//
+// Layer 2a of the e8m0 headers (#1334): the string-producing half. These build a
+// std::string, so they need <sstream> but not <iostream>; the stream operators are in
+// iostream.hpp.
+#include <string>
+#include <sstream>
+#include <iomanip>
+#include <universal/number/e8m0/core.hpp>
 #include <string>
 #include <iomanip>
 #include <universal/number/e8m0/e8m0_fwd.hpp>
@@ -45,5 +54,21 @@ namespace sw { namespace universal {
 		s << def;
 		return s.str();
 	}
+
+inline std::string to_binary(e8m0 v, bool = false) {
+	std::stringstream ss;
+	uint8_t bits = v.bits();
+	ss << "0b";
+	for (int j = 7; j >= 0; --j) {
+		ss << ((bits & (1u << j)) ? '1' : '0');
+		if (j == 4) ss << '.'; // visual separator at nibble boundary
+	}
+	return ss.str();
+}
+
+// native semantic representation: radix-2, delegates to to_binary
+inline std::string to_native(e8m0 v, bool nibbleMarker = false) {
+	return to_binary(v, nibbleMarker);
+}
 
 }} // namespace sw::universal
