+++ include/sw/universal/number/mxfloat/attributes.hpp
@@ -5,6 +5,8 @@
 // SPDX-License-Identifier: MIT
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <iomanip>       // std::setw
+#include <type_traits>   // std::is_integral_v
 #include <cstdint>
 #include <string>
 #include <sstream>
